use object single format in search
This patch fixes the object names used where checking if to use the
simple tables or the extended tables that contain tag information.

The search engine supports single and plural format at the beginning of
the search expression (i.e vm/vms cluster/clusters etc.)

The default in UI is to use the plural format but you can use also the
single format by simply typing it in teh search box
In API both formats are supported as well.

So, when searching for an object in the search text, the single format
should be used in order to match both single and plural formats.

This is true also when objects are used in the search string (i.e
Templates.<field> / template.<field>)

This fix is also related to BZ [1] which was resolved by [2] and should
be changed accordingly.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1858638
[2] https://gerrit.ovirt.org/#/c/ovirt-engine/+/110878/

Change-Id: Ia0f01724f4f016ea6a1579d85db460e99969edcb
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1989477
Signed-off-by: Eli Mesika <emesika@redhat.com>