core: invoke callback for MBS disks
When removing a VM with an MBS disk attached, RemoveVm's endAction
is not invoked after RemoveAllManagedBlockStorageDisks completed.
A consequence of this, for example, is that moving an MBS domain to
maintenance is not allowed since the command entry not being removed
from command_entities and an error is being raised:
"Error while executing action: Cannot deactivate Storage
while there are running tasks on this Storage".

The patch fixes the issue by invoking the callback for MBS domains.

Change-Id: Id85530f833963cbe9fe1024d291dfbecd5d2627f
Bug-Url: https://bugzilla.redhat.com/1959385
Signed-off-by: Shani Leviim <sleviim@redhat.com>