webadmin: remove cpu warning for HP VMs
The CPU warning on the VM list page has been shown
for all VMs whose configuration differs from the cluster's
setting. However, by selecting high performance vm
type, the cpu host-passthrough is selected
automatically and as a consequence the configuration
differs. Since it is a desired and expected state,
it was confusing to see a warning about it. The same
goes for the case when the user selects the custom cpu type.

This patch removes the cpu warning on the vm list page for
vms with the cpu host-passthrough and a custom cpu. The patch
also changes the icon on vm detail screen from warn to info
for vms with the cpu host-passthrough and the custom cpu.

Change-Id: Ib36d5e79a945fd4d9f6db9baf00850b908fd0567
Bug-Url: https://bugzilla.redhat.com/1979441
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>