api: poll the job of ImportVmFromConfiguration
Synchronous calls to ImportVmFromConfiguration would now poll for the job
of the command (PollingType.JOB) rather than its VDSM tasks. With that,
the call would resume when the command is finished and clients could safely
proceed to the next step in their execution flow withot worrying that
entities are still locked/used by the command.

Change-Id: I7673ef16ec21929c149851d3a26d64dd544a70b8
Signed-off-by: Arik Hadas <ahadas@redhat.com>