webadmin: fill in configuredCpuVerb in queries
The configuredCpuVerb is filled in the search query.
However, some of the list pages ignore the default search query
and use a different one (HostVmListModel, NetworkVmListModel)
where the configuredCpuVerb has not been filled in causing
incorrect display of a cpu warning.

The warning signs have not been visible in those tables so the
cell size was increased by 5px to be the same as in the main
vm list view.

Change-Id: I1274007168a6f085632720257f5207947974ce1b
Bug-Url: https://bugzilla.redhat.com/1971622
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>