core: Wrong allocation policy is assigned to VM imported by OVA
When importing VM from OVA and selecting Preallocated
Allocation Policy, the disks are imported as Thin
provisined.
this is because imported VM's is set to COW + Sparse
volume format and type, ignoring UI selection.

This patch takes into consideration selected
allocation policy, and converts the imported VM
disk to the right volume format if necessary.

Bug-Url: https://bugzilla.redhat.com/1986726
Change-Id: Ifab482241a134ae6387ab54c3b7833d469c4d026
Signed-off-by: Saif Abu Saleh <sabusale@redhat.com>