core: Don't hot unplug the same DIMM twice
When memory hot unplug is requested using REST API repeatedly in a
quick sequence, the same DIMMs may be selected for different hot
unplugs repeatedly.  This is because it takes some time before an DIMM
device is removed and the hot unplug memory device selection can see
it as a hot unplug candidate within that time.  It may lead to
unsuccessful hot unplugs and false assumptions about the amount of VM
memory.

Let's track DIMM devices for which a successful hot unplug has been
initiated and exclude them from future hot unplugs.  We use VmManager
for tracking the devices.

If Engine gets restarted then the list of DIMMs being hot unplugged is
lost.  Under normal circumstances, when memory hot unplugged is
configured correctly in the guest OS, it shouldn't be a big problem.
We don't store the information into the database because it's not
worth of adding a new database column.

Another possible problem with this approach may be that it's not
possible to retry a successful memory hot unplug again using REST
API (it's still possible to make it, although it should be never
needed, from the web UI).  Although this is not something that's
expected to succeed and the current code wouldn't handle it completely
correctly as for VM memory tracking, such an action was possible
before.  One notable situation where repeating the hot unplug attempt
would be needed and would make sense is after rebooting the guest OS.
Setting a next-run configuration after each hot unplug to avoid the
problem could be confusing and wouldn't be needed for a correctly
configured guests.  Let's cleanup hot unplug tracking info when VM
goes down and on guest initiated reboots.

Change-Id: I055d0a4e8599f3c4805af9db21c4af5884a03be8
Bug-Url: https://bugzilla.redhat.com/1940991
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>