core:Add console disconnect action delay
This patch adds a console disconnect action field, which lets the user
to set a delay before the console disconnect action is carried out.
Currently only Shutdown disconnect action is supported.
After a user disconnects from a vm console, if console disconnect action
is 'Shutdown', the vm now waits a specified amount of time before
shutting itself down. This prevents users from losing their progress
when accidentally disconnecting from a session.

The value of the delay field is in minutes and is sent to vdsm
whenever a vm ticket is created.

Change-Id: I6265f7489f602ed0c483cff182529f6e26035254
Bug-Url: https://bugzilla.redhat.com/1944834
Signed-off-by: Filip Januska <fjanuska@redhat.com>