core: fix command init in ImportTemplateFromConf
This bug reported NullPointerException when running Import Template
From Configuration with Quota enabled, in the validate stage, thrown
when trying to access the imageToDestinationDomainMap field. The field
is read from parameters in one of the ancestor classes, which is null
in this case.

Debug (eventually) showed that this field does get initiated in
ImportVmTemplateFromConfigurationCommand.init(), but at the validation
stage it has null value again - and it is because super.init() is
called at the end of ImportVmTemplateFromConfigurationCommand.init()
instead of in the beginning, and also because the field is initiated
in the command object, but in the parameters object it still points
to null.

Added a check if imageToDestinationDomainMap is null before
reassigning it in MoveOrCopyTemplateCommand

Change-Id: Ifc23a3a2e56154d223ab3676b0e41d3c7cb243e2
Bug-Url: https://bugzilla.redhat.com/1956295
Signed-off-by: Mark Kemel <mkemel@redhat.com>