rest-api: accelerate /api/vms?follow=snapshots
Parsing the VM configuration from an OVA that is associated with
a snapshot is a heavy task that was triggered per-VM when listing
VMs and following 'snapshots'.

This patch skips parsing the VM configuration in that case. The idea
behind it is that when snapshots are listed in this flow, the VM
configuration is rarely used (unlike when listing the snapshots for
a particular VM) and so the configuration would typically will not
be used and in case it is needed, the client can issue a call to get
the VM configuration of a praticular snapshot(s).

Change-Id: Ib364a9b24635af61c92ed1d89d31df5dce0eab7d
Bug-Url: https://bugzilla.redhat.com/1992686
Signed-off-by: Arik Hadas <ahadas@redhat.com>