core: Fix bogus add disk failure
If the underlying child command status was ENDED_SUCCESSFULLY,
TransferDiskImageCommand treated the status as failure, and failed the
upload with:

    Failed to add disk for image transfer command ...

The image transfer failure is not logged to the audit log, and the user
was left with an empty disk. If the user was lucky, booting from the
empty disk will fail.  If not, the empty disk may be discovered much
later.

Here is example log showing the absurd failure:

2021-09-06 23:52:47,446+03 INFO
[org.ovirt.engine.core.dal.dbbroker.auditloghandling.AuditLogDirector]
(EE-ManagedScheduledExecutorService-engineScheduledThreadPool-Thread-13) [] EVENT_ID:
USER_ADD_DISK_FINISHED_SUCCESS(2,021), The disk 'test.iso' was successfully added.

2021-09-06 23:52:48,456+03 ERROR
[org.ovirt.engine.core.bll.storage.disk.image.TransferDiskImageCommand]
(EE-ManagedScheduledExecutorService-engineScheduledThreadPool-Thread-87)
[96c6f66e-2fb5-4732-8352-14b27228bbd9] Failed to add disk for image transfer command
'55d551d4-3661-4326-a90a-64d3d2169eb4'

Fix the issue by treating ENDED_SUCCESSFULLY same as SUCCEEDED. Same
approach is used in HostUpgradeCallback and HostUpgradeCheckCallback.

More work is needed to fix other issue in this area:
- Disk not deleted after the bogus failure, but maybe this is not
  relevant after this fix.
- Image transfer failure is not logged to audit log.
- Unexpected status is not logged, making debugging harder.
- Code is not future proof, adding new success status will fail
  silently in runtime instead of failing to compile.

Change-Id: I916cdd618e4e456705c5494938fcd7bb2d2b5660
Bug-Url: https://bugzilla.redhat.com/1977276
Signed-off-by: Nir Soffer <nsoffer@redhat.com>