webadmin: Add a confirmation dialog to Suspend VM
This patch adds a confirmation dialog when suspending
a VM similar to the other VM power actions - shutdown,
reboot etc. Unlike the others, it is possible to dismiss
the dialog and not to show it again in the future.

If the user chooses to select "Do not show the dialog again",
this selection is stored in the user's profile in the
database. Based on this value, the dialog will or
will not appear in the future.

Change-Id: Ia05a63aa4a39cad342a57fd3394d537401999f39
Bug-Url: https://bugzilla.redhat.com/655153
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>