backup: avoid stopping backup in finalizing or final phases
Adding a validation in "StopVmBackupCommand" to verify that the backup
is still in progress (not in one of the finalizing or final phases).
1) There is no need to stop a backup that is already in one of the
finalizing phases, that can even cause a potential race.
2) Stopping an already finished backup will cause moving the backup phase
from the already final "SUCCEEDED" or "FAILED" phases to the intermediate
"FINALIZING" phase, which in turn might cause the backup to be stuck in that
intermediate "FINALIZING" phase forever, blocking another backups for that VM.
3) Locking is used to avoid simultaneous phase update.

This bug started as a more severe issue, which was mostly solved by Benny's
child command fix in 296073998c30b4da65e3bd40d2aed9496c41da1d.

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1985746
Change-Id: Ie921817bc00d4816f346b5ad34c7c50ca98bcf6b