core: reduce transaction scope for TransferImageStatus
TransferImageStatusCommand can be called from multiple different flows
and change the phase of the transfer entity. Since it is not marked as
NonTransactional, a transaction is created when it starts and committed
when it finishes.

TransferImageStatusCommand uses ImageTransferUpdater to change the
ImageTransfer entity, and the change is protected by a memory lock.
However, since TransferImageStatusCommand is currently transactional,
the actual commit of the changes to the database happens outside of the
lock, and this leaves an opportunity for a race to occur, where another
call to TransferImageStatusCommand swoops in and overrides the changes
to the entity after the lock was released but before the changes are
committed to the database.

This patch limits the scope of the transaction to the actual entity
update, to ensure the changes are committed under the protection of the
memory lock.

Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>
Change-Id: I2dc8405dea5b8cf2a7d005551f7838e1977b43f2