engine: Add warning on cluster upgrade
When upgrading cluster from a cluster level 4.5
to 4.6, the emulated machine changes as well. This
can cause problems on some Windows VMs - such as
a loss of a static IP configuration or secondary disks
being offline. These issues need to be solved from within the guest.

This patch adds warnings in case of cluster upgrade from a cluster
level <= 4.5 to the cluster level >=4.6 that some Windows virtual machines
might encounter problems. First warning lists all VMs with Windows
or Other OS that have secondary disks, the second warning all
VMs with Windows or Other OS in general. The reason is to list the
problems as specific as possible. Note that if the VM has custom
compatibility version, the warning would not be shown because
VM's emulated machine won't change.

The reason why Other is also considered is that some VMs might not be set
with the correct operating system, but with the default one (Other). While
technically the Windows VM could be set with Linux OS, this should be
considered as misconfiguration and not considered as a real possibility.

This patch also updates the cluster edit warnings dialog to
use the Warning alert.

Change-Id: I497be9a7a9a79da52ee26f470f8c90b8e8bc6d45
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1979730