webadmin: prevent multiple notificatons
Since [1] there is a special case for displaying notifications for
hosts activated from maintenance mode. Such notifications are
displayed if "end time" is after the last displayed notification.

If there is a time difference between server and the browser, the job
"end time" might be in the future. This results in displaying new toast
notifications for the same job until the browser time will advance
sufficiently.

Fixed by relying only on server time - "end time" is no longer compared
to local browser time.

[1] https://bugzilla.redhat.com/1471082

Change-Id: I1e214761875605494eb9de80f40f5de1dea940dd
Reference-Url: https://bugzilla.redhat.com/1471082
Bug-Url: https://bugzilla.redhat.com/1687845
Signed-off-by: Radoslaw Szwajkowski <rszwajko@redhat.com>