core: Add support for multiple migration connections
Libvirt allows using multiple network connections for a VM migration.
This patch adds support for the newly introduced Vdsm migration
parameter `parallel'.

Parallel migration connections are disabled by default.  They can be
enabled per cluster and per VM.  If enabled, they can specify the
number of parallel connections to use explicitly or to let Engine
select the number of parallel connections automatically, with parallel
migration connections being either required (regardless of the
bandwidth) or optional (only if there is enough bandwidth).  In the
latter case, the number is based on the migration network
bandwidth (to have enough capacity there) and the number of VM
vCPUs (to not overload the system with too many threads or to use more
CPUs than a CPU-pinned QEMU process has).  The vCPU limit applies also
when there is a user-specified number of parallel connections.

There is also a limit on the number of parallel connections accepted
by libvirt & QEMU, which is 255.

Post-copy is disabled when parallel migration connections are enabled.
This is because the current QEMU implementation of multiple migration
connections doesn't allow switching to post-copy.  It would be good to
add later some validation of this conflict.

When parallel migration connections are enabled, migration compression
is disabled because XBZRLE compression is not supported with multiple
connections.  We may enable gzip/zstd compression with multiple
connections, which is supported and should be more efficient, if
libvirt supports it.

When parallel migration connections are enabled, the available
migration bandwidth is expected to be saturated.  For this reason, we
set the maximum number of the outgoing migrations to 1.

The implementation puts information about the parallel migration type
and the custom number of parallel migrations into a single VmBase
field, to keep the number of VmBase fields as low as possible, for the
happiness of clients that query large amounts of VMs.

Parallel migrations require support on Vdsm side.  That means they are
restricted to the cluster version 4.7 or higher.

Change-Id: I201946d2e5ba9e31f3bac7e7668bd8cdc82ced61
Bug-Url: https://bugzilla.redhat.com/1975720
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>