webadmin: VmList confirmations refactoring
When there are more confirmation dialogs required
for an action, it requires to use callbacks and can
become quite unreadable and hard to maintain.

This patch introduces a helper class called ConfirmationModelChain
that takes a list of confirmations, displays them one by one
to the user and based on user action (OK / Cancel) either
calls a successCallback or just hides the confirmation dialogs.

This chain approach has been applied to the saving of a VM.

Change-Id: If0e9bf67f031e2f3e4d42a685abb1e89d91c34b8
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>