core: Fix updating timezones for existing VMs
As a part of https://gerrit.ovirt.org/115342 we have been updating
time zones for existing VMs, so time zone definitions would be matching
between Windows and non-Windows VMs. Unfortunately the script contained
an error, which removes time zone value for VMs which should not be
updated at all.
This patch fixes the script so users which haven't upgraded to 4.4.8 yet
won't be affected by this issue.
Issue for users who already upgraded to 4.4.8 will be fixed in a
different patch.

Change-Id: If5cdfcd04e03596d634b486849c25717fb7d04fe
Allow-db-upgrade-script-changes: Yes
Bug-Url: https://bugzilla.redhat.com/2005221
Bug-Url: https://bugzilla.redhat.com/2006284
Signed-off-by: Martin Perina <mperina@redhat.com>