packaging: engine-setup: Use otopi's checkForSafeUpdate
After noticing that this code was broken for a long time now, I moved it
to otopi, adding there automation checks, and with this patch use the
new code. Requires a patch recently merged to otopi [1], thus updating
the spec file.

Also make the output to the user slightly more informative:
- output 'No update for Setup found' right after checking that
- output 'Checking for product updates' also if [2] is in the answer
file or in '--otopi-environment'.

[1] https://github.com/oVirt/otopi/pull/11

[2] OSETUP_RPMDISTRO/enableUpgrade=bool:True

Change-Id: Ieaf51572ba2d16ca67c8b23f7b950d56772e2f15
Signed-off-by: Yedidyah Bar David <didi@redhat.com>