webadmin: add a warning on removing a Storage Domain with volume leases
As a part of decommissioning a storage domain, users migrates all the disks
and removed the SD from the engine, all volumes migrated automatically
beside the ones with a lease that needs to be migrated manually.
Currently, the deletion operation passes successfully without any warning
in the UI and only with a warning in the events which get unnoticed.

This patch adds a warning in that cases at the removing storage domain or
data center windows to make sure the user will be aware that there are still
volumes on the SD and they need to be moved.

Screenshot: https://imgur.com/a/T72wY39

Change-Id: Ic7b2d4a79517c92a8fe8f23f7228928ce7dd6b49
Bug-Url: https://bugzilla.redhat.com/1979797
Signed-off-by: Bella Khizgiyaev <bkhizgiy@redhat.com>