core: extract_ova: wait for loop device
When creating a loop device, losetup command may fail temporarily.
Because of that we need to retry the command up to 10 seconds to get the
loop device available.

The issue and the fix in kernel-utils is described in
https://www.spinics.net/lists/kernel/msg3977449.html. We should try to
solve it on our end for users who didn't get that fix yet.

Change-Id: I52b5b67a267d190140c83a9c21f2f7a10a44e898
Bug-Url: https://bugzilla.redhat.com/2001565
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>