core: Allow placing domain in maintenance mode with suspended VM
A storage domain cannot be set to maintenance if there is
a suspended VM with a disk that resides on that storage domain.

There is nothing that prevents for setting the storage domain
to maintenance at this case.

This patch allows setting the domain to maintenance even if there are
suspended VMs.

P.S - the suspended VMs cannot run as long as the storage domain in
maintenance mode.

Change-Id: Ie8d2f25523c4f20d73ebde57e44980087e4f2c54
Bug-Url: https://bugzilla.redhat.com/1838089
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>