core: release memory locks before unlocking entities when image transfer is finished
Memory locks are released after the custom "endAction()" code
("endSuccessfully()" / "endWithFailure()") where entities are unlocked,
see [1].
This is a problem for clients trying to wait for the completion
of an async operation, since they might try to perform another
image transfer operation before memory locks were released and fail with a
"Disk is locked" error.

By releasing the memory locks, of which the status is only known
internally, *before* unlocking of the DB entities, we allow clients
to wait only on the entity's status (disk status, transfer status, etc).

[1] https://github.com/oVirt/ovirt-engine/blob/ebcf3cfd4bb1dccd0c136f360135ee080f5d634c/backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/CommandBase.java#L524

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1923178
Change-Id: Id6189296b97451839aeb89502ee8924beb7b4a03