core: release memory locks before unlocking entities when backup is finished
Memory locks are released after the custom "endAction()" code
("endSuccessfully()" / "endWithFailure()") where entities are unlocked,
see [1].
This is a problem for clients trying to wait for the completion
of an async operation, since they might try to perform another
backup operation before memory locks were released and fail with a
"Disk is locked" or "The VM is during a backup operation" errors.

By releasing the memory locks, of which the status is only known
internally, *before* unlocking of the DB entities, we allow clients
to wait only on the entity's status (disk status, backup status, etc).

[1] https://github.com/oVirt/ovirt-engine/blob/ebcf3cfd4bb1dccd0c136f360135ee080f5d634c/backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/CommandBase.java#L524

Signed-off-by: Pavel Bar <pbar@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1984308
Change-Id: I8d4c5f23a113eb8d1120bb0fe5084d950c309580