engine: Fix engine crash when there is VM with cloud-init network
When adding statless VM that contains cloud-init network
configuration and running it, the engine crashes after
restart, the reason for that is we were using streams parallel,
which made threads run with wrong class loader.
this caused the engine to fail to deserialize VmInitNetwork
class because the class was not found in the running thread's
class loader.

This patch fixes this issue by using ThreadPoolUtil threads
to retrieve the VmInitNetwork configurations in parallel,
which uses the application class loader.

Bug-Url: https://bugzilla.redhat.com/2000364
Change-Id: I0136c27b6d1d2ce58ca62fdb96f0555907990038
Signed-off-by: Saif Abu Saleh <sabusale@redhat.com>