engine: watchdog service for host monitoring
This patch adds additional layer of monitoring on top of VdsManager and
HostMonitoring.

The motivation behind this change is a difficulty to investigate and
debug very rare production issues causing engine to lost connectivity
with hosts. In a result - these hosts end up in illegal states like
'Unassigned' or simply cannot be activated.

Based on configuration described below the service occasionally prints
information about hosts that were not 'refreshed' by HostMonitoring for
given period.

The watchdog service is configured by optional configuration properties:

$ engine-config -g HostMonitoringWatchdogIntervalInSeconds
HostMonitoringWatchdogIntervalInSeconds: 900 version: general

$ engine-config -g HostMonitoringWatchdogInactivityThresholdInSeconds
HostMonitoringWatchdogInactivityThresholdInSeconds: 1200 version: general

In order to disable watchdog service set:
$ engine-config -s HostMonitoringWatchdogIntervalInSeconds=0
and restart ovirt-engine

Please note that this is the initial version that explores this area and
in future additional information can be added to the log statements or
the conditions for raising warnings could be changed according to the
needs

Change-Id: I83c63568dbbd2cda371952407ccf2d57a581e6e9
Bug-Url: https://bugzilla.redhat.com/1897114
Signed-off-by: Artur Socha <asocha@redhat.com>