core: avoid passing lock to child command in StartVmBackup
By default when invoking child commands, the context of the parent is
cloned along with the parent's locks and passed down the child.
This can cause the lock to be prematurely released by the child command,
making the parent continue its execution without a memory lock.
Possibly allowing other operations to run on the VM or the disk during
backup (RemoveDisk for example).

This patch avoids the passing of the lock by using
`cloneContext().withoutLock()`

This is similar in nature to dd7e22f:
    core: avoid passing lock to child command for image transfer

Additionally, finalizeBackup has to be reordered to ensure the lock is
released before RemoveScratchDisk is called, otherwise it will fail to
acquire the lock.

Change-Id: I0d72cc8cd197311dcf9cb06e2685c7b81f67babb
Bug-Url: https://bugzilla.redhat.com/2015470
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>