rest: Add 'include_template/active' param to GET disks/disksnapshots
Following change https://gerrit.ovirt.org/c/ovirt-engine/+/115664
Provide 'include_active' and 'include_template' flags also for
disks/xxx/disksnapshots.

  GET .../api/disks/xxx/disksnapshots
Returns:
  disk_snapshot: id=bbb parent=aaa
  disk_snapshot: id=aaa parent=zzz

  GET .../api/disks/xxx/disksnapshots?include_active
Returns:
  disk_snapshot: id=ccc parent=bbb (Active)
  disk_snapshot: id=bbb parent=aaa
  disk_snapshot: id=aaa parent=zzz

  GET .../api/disks/xxx/disksnapshots?include_active&include_template
Returns:
  disk_snapshot: id=ccc parent=bbb (Active)
  disk_snapshot: id=bbb parent=aaa
  disk_snapshot: id=aaa parent=zzz
  disk_snapshot: id=zzz            (Template)

When 'include_active' flag value is true, the result returned for
the request will include disk snapshots and also the Active VM
snapshot.

When 'include_template' flag value is true, the result returned for
the request will include disk snapshots and template disk image, if
it is the parent of one of them. These might be required to
reconstruct the snapshot chain in case the VM was created from
template with Thin Storage Allocation selected, and a template disk
image being a parent of one of the snapshots and not returned without
the 'include_template' flag.

Change-Id: I96f408089dae8fc21337138f1ac0bac474c6b329
Bug-Url: https://bugzilla.redhat.com/1900597
Signed-off-by: Mark Kemel <mkemel@redhat.com>