core: Add newly available Hyper-V flags
Newer versions of QEMU/libvirt add support for additional Hyper-V
flags: hv_stimer_direct, hv_ipi, hv_evmcs.  As discussed in the
attached bug, all those flags can be safely added.  The only
restriction is that hv-evmcs can be enabled only on Intel CPUs, it
cannot be used on AMD.

However VMs with hv-evmcs cannot be started on hosts without nested
virtualization enabled.  Having the flag is not critical, QEMU
documentation even says: "Some virtualization features (e.g. Posted
Interrupts) are disabled when hv-evmcs is enabled. It may make sense
to measure your nested workload with and without the feature to find
out if enabling it is beneficial."  For these reasons, we do not
enable the flag.

The new flags have been available in the platform since cluster level
4.5.  But we enable them only for the future cluster level so that
users can stay on 4.6 in case the flags cause any trouble.  The added
flags are needed only for nested virtualization or some optimizations
so they are not going to be missed much in supported flows.

Change-Id: Ic037ac7c3259aae6083dfc264073499fab927a33
Bug-Url: https://bugzilla.redhat.com/2003862
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>