core: set correct host for removing a backup scratch disk
When VM stopped during a backup, finalizing the backup will
not teardown and remove the created scratch disks.

Set the host for scratch disk removal according to the VM
backup and not according to the host that runs the VM which
is none when the VM is down.

Change-Id: I19552b80c8c9deee7c4362c2991392d5c06f3b39
Bug-Url: https://bugzilla.redhat.com/2013932
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>
(cherry picked from commit 4c9cd925622e699104081495fd756211adb3eebc)