core: fix NPE in get disk snapshots query
Fixing a bug introduced in the 'include_template' for SD snapshots
patch: when the api/storagedomains/xxx/disksnapshots?include_active
request is sent and there is an ISO disk in the storage domain -
NullPointerException is thrown. That is because ISO disk image object
has null value in the vmEntityType field.

Change-Id: I56054f2e4e89fd65e0cb46ab7e1d1ae7e7c22cd2
Bug-Url: https://bugzilla.redhat.com/1900597
Signed-off-by: Mark Kemel <mkemel@redhat.com>