core: allow configurable backup scratch disk size
Setting the scratch disk initial size was done according to
the disk format and the storage domain it is based on.

 - For COW block-based disks the initial size sets according to
   the disk actual size fetched from VDSM.

 - Otherwise, the initial size sets to - 0.

This is wrong, disk initial should be set as folows:

  - File-based scratch disk: initial size should be - '0'
    (which mapped to None in VDSM), since initial size
    isn't supported for file-based storage domains.

  - Block-based scratch disk:
      * Backed-up disk format is RAW: the initial size
        should be the backed-up disk actual size.

      * Backed-up disk format is COW: the initial size
        should be measured from the entire chain.

When starting a live VM backup for a VM with disks on
block-based domain, the size of the created scratch disks
will be equal to the size of the theoretical amount of data
the guest may write to the scratch disk, if the guest
overwrites the entire disk data during the backup.

When there is limited space on the block-based domain,
the creation of the scratch disks will fail and only a cold
backup can be taken for the VM which does not require
scratch disks creation.

By allowing to configure the created scratch disk size,
scratch disks can be created according to the given percentage
from the size of the backed-up disk, which will make a live backup
possible for large block-based disks with a limited capacity of
storage on the storage domain, the minimum size for a scratch disk
can is set to 4GB by default but can be changed by setting the
MinBackupBlockScratchDiskInitialSizeInGB configuration value on the
engine config.

Change-Id: I4d44597e530242648d34f490d6591a7d7a9326f9
Bug-Url: https://bugzilla.redhat.com/2018971
Bug-Url: https://bugzilla.redhat.com/2018986
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>