core: lock disks at the start of VM backup
Currently, the VM disks are lock in the database fairly late after
backup is started.
This patch locks the disks at the start of StartVmBackupCommand#execute
as it's better to reflect the disks are in-use sooner rather than later
and adds another line of defence in case another operation does not
properly takes memory locks.

Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>
Change-Id: Ic82c0f4d072078448453c913c1058de0f4c4af22