core: Enable IOMMU caching_mode when mdev devices are present
When the maximum number of CPUs is high enough, we add an IOMMU
device.  When a vfio-pci device is present, caching mode must be set
for the IOMMU device, see https://wiki.qemu.org/Features/VT-d.  When
it is not set, the VM may fail to start.  This can happen e.g. for VMs
with 1 CPU socket and 16 cores (and thus 256 max CPUs).

This patch fixes the caching mode omission and adds the corresponding
option to the domain XML.

Change-Id: I2ae85222bd90088f8fd75715d7e84ecc6fe9ddff
Bug-Url: https://bugzilla.redhat.com/2013752
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>