webadmin: Create migrations builder common for vm and cluster
Some parts of copying data from ClusterModel to Cluster entity
and from UnitVmModel to VmBase (and vice versa) could be shared.
Since we already have a common interface HasMigrationOptions,
this patch adds the corrensponding model interface and a builder
that copies data from one to another. The builder is then used
for both, clusters and vms.

Change-Id: Ic98fd35cc9623de3fa7466484ed93d609a8b353d
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>