webadmin: Refactor model to entity builders
There are 4 usages for UnitVmModel to VmBase
builders - for VMs, Templates (new and edit
use different builders) and Pools.

This patch removes classes MultiQueuesVmBaseBuilder
and CpuProfileUnitToVmBaseBuilder because they
were used in all scenarios, contained only one line
and should therefore be part of the Core builder common
for all scenarios.

This patch also adds PoolUnitToVmBaseBuilder used specifically
for the pools. This helps to see the main difference compared
to the FullUnitToVmBaseBuilder (usage of Core instead of Common
- that means leaving out quota and numa configuration).

Change-Id: I43e711711f8de7f231895a7c75cf71c6f8fbc5d1
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>