frontend: fix fips kernel args
When we have the boot=UUID=<ID> in the kernel args we set the value of
the kernel boot UUID to empty string. But, later on we check only if
it's null. In such case we will end up having: "fips=1 boot=UUId="
without any ID, which makes a fault on boot when the user is not
cautious. In such case we need to append only "fips=1".

Change-Id: I1b482b2642bf479ed1cca0c38c9c707a56199d9e
Bug-Url: https://bugzilla.redhat.com/2013430
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>