core: persist VM backup type (live/cold)
VM backup can be taken only when the VM is in 'UP' or 'DOWN' states.
When the VM is down, a cold backup will be taken for it, when it is
running a live backup is taken.

But VM can enter to 'PAUSED' or 'UNKNOWN' state during different phases
of the live backup operation, due to a lack of storage space for example,
in those cases, the system needs to continue the regular live backup
flow and not treat it as a cold backup.

By adding a new field to the backup object to mark if the backup is live
or cold, the system can tell the type of the backup without relying on
the VM state which is dynamic and fragile.

Change-Id: I046c5b0278c3fcd10c892d001c9fa6248b9303b1
Bug-Url: https://bugzilla.redhat.com/2016913
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>
Signed-off-by: Martin Perina <mperina@redhat.com>