webadmin: Disable fields for VMs in pool
VMs in the pool should have (almost)
all fields disabled when opening its edit dialog.
This patch adds the missing fields to the
setVmAttachedToPool() method in the UnitVmModel which
is responsible for disabling them.

Also, the patch overrides setChangeable(Boolean, String)
method in the NotChangableForVmInPool* classes as
it was missing and it was possible to set the field
as changeable using this method.

Change-Id: Ib1dd1f8c55195fa8f0d9dba3ba074542a53baff0
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>