core: change tools warning to agent
This patch changes the UI warning for Windows VMs when they have out of
date guest agent within the guest. Before moving to VirtIO-Win, we used
the Windows guest tools, the warning sign was about it. Now, we check
only the guest-agent. Therefore, the naming changed in this patch.
Also, to ease on debugging, a log added, telling what is the available
version by the engine and what the VM has.

Change-Id: I33d57bae85906a881b03c74f2940a12a29932cf0
Bug-Url: https://bugzilla.redhat.com/1940494
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>