core: align network stats tx\rx_drops definition & type
rx_drop and tx_drop values reported by vdsm are defined as 'Number of
packets received but not processed', aka -dropped [1][2]. This should be
reflected by their names in engine.
As they are of type U64 (unsigned long) [1], they are whole numbers. This
should be reflected by the type of their counterparts in engine.

[1] https://www.kernel.org/doc/html/latest/networking/
    statistics.html#c.rtnl_link_stats64
[2] https://github.com/oVirt/vdsm/blob/
    51ee47b235ec28744165190a22e211d43fb371b7/lib/vdsm/api/
    vdsm-api.yml#L5076

Change-Id: Iaeed9e91b0f051fdf52f8c974d5d707028d2abe0
Bug-Url: https://bugzilla.redhat.com/1980192
Signed-off-by: Eitan Raviv <eraviv@redhat.com>