backend: host maintenance incoming migration
When triggering a host maintenance it will check the incoming migrations
and cancel them. It might happen that an incoming VM finishes the
migration when we already see it as incoming in the maintenance command
and triggering the cancel migration command. In such case, the command
will fail and as a result the maintenance command as well.

In the patch, once we will fail to cancel the migration we will wait 15
seconds to get the DB refreshed and continue with the maintenance
command.

Change-Id: If79aa9c6fb606ef64718e0426a248a35b80bc9a7
Bug-Url: https://bugzilla.redhat.com/1887174
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>