fix endless loop in soft fencing
In case that a host in non responding due to a miscommunication between
the vdsm running on that host and the engine machine, the host will be
first soft fenced by restarting its vdsmd service, however since the
communication with the engine machine stayed blocked, this will not
solve the issue and the host is supposed to be rebooted if it has PM
configured.
However, a bug in the code cause to wait forever to the host to be up in
soft fencing while not incrementing the attempts number resulted actually
in an endless waiting loop for host UP status which will never occur in that case.

The code calls isHostInGracePeriod from two scenarios, one is from the
VdsManager::handleNetworkException, in this code the isHostInGracePeriod
method is called with sshSoftFencingExecuted set to false which does not
double the number of attempts, the other call is from
SshSoftFencingCommand::checkIfHostBecomeUp method in which
isHostInGracePeriod is called periodically with sshSoftFencingExecuted
set to true, but nothing increments the attempts in the while loop,
causing it to wait forever to host status to become UP.

This patch increments the attempts when network exceptiion occurs in any
case, preventing the endless loop to occur.

This patch also stops the SSH waiting loop wither if number of
attempts reached or timeout exceeds , the sooner takes.

Change-Id: Ie4a3d10327fee09f928b812b6819fc6069e0c4fc
Bug-Url: https://bugzilla.redhat.com/2007286
Signed-off-by: emesika <emesika@redhat.com>